EMCPSSF-404 - Data cannot be retrieved due to an unexpected error. Precondition Failed

+++ savedsearch-web/src/main/java/oracle/sysman/emSDK/emaas/platform/savedsearch/ws/rest/search/SearchAPI.java
@@ -333,6 +333,9 @@ public class SearchAPI
 		SearchManager sman = SearchManager.getInstance();
 		
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			odsService.deleteOdsEntity(searchId);
 			EmAnalyticsSearch eas = sman.deleteSearch(searchId, false);
 			// TODO: when merging with ZDT, this deletionTime should be from the APIGW request
@@ -364,11 +367,14 @@ public class SearchAPI
 			return Response.status(Response.Status.BAD_REQUEST).entity("The param isExactly is invalid").build();
 		}
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			LOGGER.debug("Calling searchManager.deleteSearchByName");
 			searchManager.deleteSearchByName(name, Boolean.valueOf(isExactly));
 		} catch (EMAnalyticsFwkException e) {
 			LOGGER.error(e.getLocalizedMessage());
-			return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+			return Response.status(e.getStatusCode()).entity(e.getMessage()).build();
 		}
 		return Response.status(Response.Status.OK).build();
 	}
@@ -492,6 +498,9 @@ public class SearchAPI
 		Search searchObj;
 		SearchManager sman = SearchManager.getInstance();
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			if (updateCategory != null && "ORACLE_INTERNAL".equals(updateCategory)) {
 				searchObj = createSearchObjectForEdit(inputJsonObj, sman.getSearch(searchId), true);
 
@@ -530,6 +539,9 @@ public class SearchAPI
 		Search savedSearch = null;
 		
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			savedSearch = sman.getSearch(searchId);
 		} catch (EMAnalyticsFwkException e) {
 			message = e.getMessage();
@@ -586,6 +598,9 @@ public class SearchAPI
 		// find and return the whole object
 		SearchManager sman = SearchManager.getInstance();
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			Search searchObj = sman.getSearchWithoutOwner(searchId);
 			jsonObj = EntityJsonUtil.getFullSearchJsonObj(uri.getBaseUri(), searchObj);
 			message = jsonObj.toString();
@@ -744,7 +759,9 @@ public class SearchAPI
 		JSONObject jsonObj;
 		SearchManager sman = SearchManager.getInstance();
 		try {
-
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			Search searchObj = sman.getSearchWithoutOwner(searchid);
 			String[] pathArray = null;
 			if (bPath) {
@@ -770,6 +787,9 @@ public class SearchAPI
 		String message = null;
 		int statusCode = 200;
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			SearchManager sman = SearchManager.getInstance();
 			java.util.Date date = sman.modifyLastAccessDate(searchId);
 			message = String.valueOf(DateUtil.getDateFormatter().format(date));
@@ -777,7 +797,7 @@ public class SearchAPI
 		catch (EMAnalyticsFwkException e) {
 			LOGGER.error(e.getLocalizedMessage());
 			message = e.getMessage();
-			statusCode = 404;
+			statusCode = e.getStatusCode();
 		}
 		return Response.status(statusCode).entity(message).build();
 	}
@@ -793,6 +813,9 @@ public class SearchAPI
 		String version;
 		String rel;
 		try{
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavilException();
+			}
 			Search search = searchManager.getSearch(searchid);
 			if(search.getCategoryId()==null){
 				return Response.status(Response.Status.NOT_FOUND).entity("Search with id "+searchid
@@ -821,7 +844,7 @@ public class SearchAPI
 		}catch (EMAnalyticsFwkException e) {
 			LOGGER.error((TenantContext.getContext() != null ? TenantContext.getContext().toString() : "") + e.getMessage(),
 					e.getStatusCode());
-			return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+			return Response.status(e.getStatusCode()).entity(e.getMessage()).build();
 		}
 	}
 	private Search createSearchObjectForAdd(JSONObject json) throws EMAnalyticsWSException
